Fine-grained compatibility definitions
Turns out that you were on a good track with commit 5ec8969, and what
I did with ec6f461 was nonsense.

`LUA_COMPAT_ALL` and `LUA_COMPAT_5_1` are used when compiling Lua, to
add compatility functions to liblua, and don't have the desired effect
here. If compat functions are not available on the system (apparently
neither Debian nor Homebrew enable them), then we need to define the
compatibility functions ourselves.

With this commit, lua-zip can now be used on Lua 5.1, 5.2, and 5.3,
even if compatibility functions are not present in the Lua library.